8343137: C2: VerifyLoopOptimizations fails with "Was reachable in only one"
